Remove mysql option from postgres connect string
This has been present in the URL since the "initial commit" (79918e5),
where there was a single function used for both MySQL and Postgres.
Since then, the code has been refactored so that there is a separate URL
generation function for MySQL and Postgres, however this option has
remained.

This is not an option that's used by Postgresql[1].  Some client
libraries (eg the libpq C library) will emit a warning with this option
present.  Others (eg the Go github.com/lib/pq library) will error.

The result of this is that any client using the VCAP provided
credentials to connect to the postgres database can fail to parse the
connection string and throw and error - previously we have worked around
this by stripping the parameter before using it, but that is not an
acceptable solution for tenants in production.

[1]http://www.postgresql.org/docs/9.5/static/libpq-connect.html#LIBPQ-PARAMKEYWORDS