Disable RDS backups by default (workaround)
Some facts:

* We wish to have the ability to disable RDS backups.

* Disabling RDS backups is achieved by setting the backup retention
  period to 0.

* The golang zero value for an Int is 0.

* The AWS API bindings default for backup retention is 1.

* The RDS broker allows you to override the backup retention period as
  long as the value is > 0.

Hence, we cannot turn off backups either from a plan or with
user-specified parameters.

We decided that having backups default to being off (0) is reasonable
as the plans themselves can override this - we may even wish to expose a
postgres plan that explicitly does not have backups for testing.

This change simply takes whatever value is defined in the plan rather
than requiring it to be > 0, and the plan default is 0 (off).