Make some test strings constants.
These are currently variables, but aren't modified by any of the tests.
Additionally, if any test does modify them, that state will leak out
into other tests. this is because they are initialised when the test id
defined, instead of in a BeforeEach block.

Making these constants makes this clearer, and prevents anything
modifying these which would lead to unpredictable results.