Allow dbPrefix to include - and _
DB instance identifiers can't contain an _, and db names can't contain a
-. Currently generated uuid's have - and _ replaced with the other when
generating the instance ids and db names. This replacement isn't applied
to the dbPrefix, meaning that if it contains either of these characters,
the broker will error when provisioning.

This applies the same replacement logic to the dbPrefix.