Fix for allowing backup_retention_period to be 0
56edf80 Allowed setting the retention period to zero in the broker. It
missed out this extra change in the RDS interface, which also had a
check for > 0. This resulted in the backup_retention_period being set to
`null` in the request to AWS, and the AWS default of 1 day being
applied.

The value in the request to AWS is a null instead of 0 because the
`rds.CreateDBInstanceInput` type defines its `BackupRetentionPeriod`
field as `*int64` instead of `int64`, meaning that the zero value for
this field is nil (which translates to a null in JSON).