Fix the usage of md5 in utils.GetMD5B64
The 2 functions `md5.Sum` and `hash.Hash.Sum` are not equivelent.
`md5.Sum` takes an byte slice, and returns its MD5 sum[1]. The Hash type
(returned from `md5.New`) on the other hand has a different
interface[2]. It implements io.Writer, and therefore expects data to be
added to the hasher via a Write call. Its Sum function will append the
hash to the given buffer.

This means that the previous implementation was appending the hash of
empty string to the given text, and then truncating the result to the
requested length, meaning that it was effectively base64 encoding the
text without MD5 hashing it first.

There was also an issue with the truncation where it was being truncated
based on the length of the output from the md5 function, not based on
the output of the base64 function.

[1] https://golang.org/pkg/crypto/md5/#Sum
[2] https://golang.org/pkg/hash/#Hash