Prefix master password generation with a seed
Presently, if you know the service GUID for a database instance, you can
determine the master password. This should not be possible as the
password is supposed to be secret.

This adds a configuration variable `master_password_seed` which is added
to the instance_id before hashing. This means that it's no longer
possible to determine the master password from the instance_id alone.