Ensure generated usernames and passwords are valid.
RDS master passwords can't contain '/', '"' or '@' [1]. Postgres
usernames have to start with a letter, and can't contain a '-'.[2]

This switches the base64 encoding to use the URL safe scheme, which uses
'-' and '_' instead of '+' and '/'.

To ensure the usernames are valid, they are prefixed with a 'u', and
have '-' replaced with '_'.

[1]http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/CHAP_Limits.html#RDS_Limits.Constraints
[2]https://www.postgresql.org/docs/9.5/static/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS