Remove support for bind-specific db name.
We're not using this feature at the moment (bind parameters are
disabled), and it adds a fair amount of complexity to the code. This
will also make the upcoming changes to how permissions work much more
complex as well. It therefore makes sense to remove this to reduce the
maintenance burden.

Additionally, now that we're no longer supporting bind-specific
databases, we need far fewer methods from the SQLEngine. We only need to
create and drop users, and grant privileges to a user.