Fix ginkgo install error in Travis build
Now that godep is being more selective about what's vendored, the ginkgo
cli tool is no longer being included - it's not referenced by the code
anywhere. This meant that the `go install` call was failing because it
couldn't find the package.

We're not actually using any features from the ginkgo cli, so we can
just use `go test` instead, and avoid this problem.