Re-use same user for all postgres binds.
Previously the PostgreSQL engine copied the MySQL engine and granted all
rights on the database to the newly created user. In PostgreSQL this
will give the user rights to create tables, but because it has a more
finely-grained permission model than MySQL this does not give the user
rights on existing tables, or tables newly created by other users.

Only the owner of a table can alter/drop it, and you cannot grant this
permission to other users. Users who are the owners of tables cannot be
removed until the table ownership is changed.

We solce these issues by using the same owner user for all
binds that have permission to make schema modifications (at present,
that's all binds). This means that we need to be able to return the
password for this user in all bind calls, but the password is only
known at the point the user is first created, so this isn't possible.
It's therefore necessary to store the password for this user in a
dedicated database within the RDS instance.