Add functions to symetrically encrypt/decrypt
These will be used to encrypt passwords before storing them in the
database.

These have been implemented using the AES256 bolck cipher configured in
Galois/Counter Mode (GCM)[1]. This is an authenticated encryption[2] mode
of operation that includes a message authentication code so that it's
possible to verify that the data is decrypted correctly.

We've added a test that encrypting the same string twice results in
different value because it's important that symmetric encryption uses an
IV or nonce so that it's not possible to obtain information about the
key by comparing 2 or more different ciphertexts.

For a good summary of thes encryption modes etc, see
http://stackoverflow.com/a/22958889/983841

[1]https://en.wikipedia.org/wiki/Galois/Counter_Mode
[2]https://en.wikipedia.org/wiki/Authenticated_encryption