Encrypt passwords stored in the database.
In order to reduce the risk of passwords being leaked, we want to
encrypt all passwords that we store in the state database. These need to
be encrypted, and not merely hashed because we need to be able to
decrypt the passwords so that we can return them in Bind calls.