Ping the connection inside Postgres engine Open()
Open() may not actually connect to the database so we are not sure
the connection details are right:
https://golang.org/pkg/database/sql/#Open

It is an issue as we rely on it to validate the connection details.
We now ping the connection to validate properly.

A new generic error LoginFailedError is returned in case of login
failure.