Fix broker resetting with incorrect master password.
The routine that updates the master passwords if the seed has changed
had a problem where it would occasionally use the wrong password,
meaning that the broker could no longer access the instance.

This is caused because the function used to determine the CF instanceID
from the RDS instanceID was stripping too much from the front of the
dbInstanceID. `strings.TrimLeft` takes a character set as its second
argument[1], and strips from the left all characters in that set up to the
first character that's not in the set. This meant that if some of the
starting characters in the CF instanceID were included in the dbPrefix,
they'd get stripped as well.

Replacing this with `strings.TrimPrefix`[2] does what's indended. This
also exposed another bug where the handling of '_' in the dbPrefix was
missing. This needed to be addressed with a `Replace` call.

[1] https://golang.org/pkg/strings/#TrimLeft
[2] https://golang.org/pkg/strings/#TrimPrefix