Increase the RDS creation timeout
While RDS instances take <10mins to create once the creation starts,
we're seeing the creation itself not start for quite some time after
it's requested. This is leading to the the tests failing and orphaned
RDS instances being left behind.

Doubling the creation time should hopefully avoid this problem, but we
should also query this with AWS themselves as we're frequently seeing
extended creation times.