Treat databases that no longer exist as deleted
Previously if an RDS instance was already deleted and the user attempted
to delete it, it would cause the broker to no longer allow you to create
or delete any other database.

This change treats a database being missing (either because we removed
it or RDS did) as a successful deletion, because it already has been.