Add small and large instance sizes
In order to meet user needs, we'd like to allow them to have a choice of
two additional RDS Postgres instance sizes.

We've added the `db.t2.small` and `db.m4.2xlarge`, along with
appropriate description and metadata. Extended and amended two sets of
tests, in order to cover the new instance choices.

We have also covered a common request of mentioning the actual instance
name, in the description.