Fix async delete for updated brokerapi library.
We hit an issue with async deletion of service instances following the
switch of the brokerapi from the frodenas fork to the pivotal-cf
version. The frodenas version returns a 410 Gone if the broker
implementation returns an ErrInstanceDoesNotExist[1], whereas the
Pivotal-cf version returns a 404 Not Found[2]. The service broker API
docs specify that a 410 can be returned to indicate to CF that service
deletion has completed, and this is how this was working previously. A
404 is treated as an error, so CF keeps on polling for deletion to
complete.

This therefore updates the broker to return a succeeded response to the
lastOperation polling when DB instance has gone. In order to only do
this in the case of delete requests, we return a "deprovision" string
in the operation field from the Deprovision call. Cloudfoundry provides
this string to the lastOperation call[3] so we can test for it.

[1]https://github.com/frodenas/brokerapi/blob/e43c1f9/api.go#L368-L370
[2]https://github.com/pivotal-cf/brokerapi/blob/bd7932b/api.go#L334-L338
[3]https://docs.cloudfoundry.org/services/api.html#polling