Revert "Merge pull request #18 from jmcarp/update-deps"
We've run into a number of issues with the updated code:

Async delete no longer works correctly due to changes in the brokerapi
library. The frodenas version returns a 410 Gone if the broker
implementation returns an ErrInstanceDoesNotExist[1], whereas the
Pivotal-cf version returns a 404 Not Found[2]. The service broker API
docs specify that a 410 can be returned to indicate to CF that service
deletion has completed, and this is how this was working previously. A
404 is treated as an error, so CF keeps on polling for deletion to
complete.

The Bind call now returns a different type for the Credentials hash,
which no longer serializes to the same JSON document. Specifically, the
field names are no longer lowercase, which means that several buildpacks
no longer populate the DATABASE_URL environment variable.

This reverts commit 1b3cec60663f5daf365b15810063b90dd133746c, reversing
changes made to a34230577537971cda01fc83ce0d649b29ebabb0.

[1]https://github.com/frodenas/brokerapi/blob/e43c1f9/api.go#L368-L370
[2]https://github.com/pivotal-cf/brokerapi/blob/bd7932b/api.go#L334-L338