Manage postgres users with event triggers.
Note: comment from original commit b6c6588c2e406347a10c423b7c06dc65a13e58be
      by @alext and @jonty

Previously the PostgreSQL engine copied the MySQL engine and granted
all rights on the database to the newly created user.

In PostgreSQL this will give the user rights to create tables, but because it
has a more finely-grained permission model than MySQL this does not
give the user rights on existing tables, or tables newly created by other
users. Only the owner of a table can alter/drop it, and you cannot grant this
permission to other users. Users who are the owners of tables cannot
be removed until the table ownership is changed.

Previously we added logic to always return the same user for every binding,
but in this commit we solve all of these problems by creating a "parent"
group that is given ownership of all tables, sequences, and functions.
New users are granted rights on this group and thus can carry out all
owner actions on group-owned items.

Unfortunately there is no way to set a default owner on newly created items (it
defaults to the user that created them), so we install a DDL event trigger
(https://www.postgresql.org/docs/9.5/static/event-trigger-definition.html) that
fires when every `CREATE`, `ALTER`, `DROP`, `GRANT` or `REVOKE` command
finishes. It simply uses the built-in `REASSIGN OWNED BY` command to hand over
ownership of all resources owned by the current user to the parent
`$DBNAME_owner` group.

We use a per-database group name because roles/users/groups are global across
the PostgreSQL instance, and we may wish to support more than one database per
PostgreSQL instance in the future.

We also tried the following approaches:

* Using `ALTER DEFAULT PRIVILEGES`
  (https://www.postgresql.org/docs/9.5/static/sql-alterdefaultprivileges.html) on
  every newly created user to `GRANT ALL ON PUBLIC` - this means that every user
  can `SELECT`, `INSERT`, and `DELETE`, but because only the table owner can
  `ALTER` or `DROP` this will not allow other bound users to run migrations. This
  is also limited to the `PUBLIC` (default) schema, so would fail to work for any
  applications that have custom schemas.

* Making the group the owner of the `PUBLIC` schema. This allowed members of
  the group to `DROP` tables within the schema, but still did not allow them to
  `ALTER` these tables.