Provision() can restore from snapshot
User can pass a parameter to pass a GUID to restore from snapshot, using
the parameter `restore_from_latest_snapshot_of`:

This parameter signifies two things:
 1) that you are creating an instance from a snapshot, and
 2) the Cloud Foundry instance ID of the service you are restoring from.

This name makes it clearer we are using the latest snapshot and cleaerer
what the client has to pass in.

In this case we call RDSDBInstance.Restore(), which tests are really
similar to the Create(), mainly a subset. We expect the new instance to
contain a tag `Restored From Snapshot`.

We add logic to list the snapshots from RDS, and return them in order in
the RDSDBInstance.DescribeSnapshots(). We want to use the latest snapshot
to recover from, so we pick the first one assuming that they are returned
in order.

We add logic to get the Snapshots tags, as we will use the snapshot tags
to check that the snapshot was generated in the right Organization, Space
and with the right plan. For that we query and return the tags of all the
snapshots when calling RDSDBInstance.DescribeSnapshots.

Using these tags we check:

 * Restore from snapshot must be in same space/org

    Check that the instance being created has the same Organization and
    Space than the snapshot that is being restored from.

    We want this restriction to avoid leak data between spaces, and
    prevent users without access to some space to restoring
    any instance from that space.

    Note: We might improve this logic later, providing capabilities to tag
    instances to allow restore between spaces or even organisations.

 * Restore snapshot must be in the same Plan

   Plans might have incompatible properties: storage size, encription, etc.

   To ensure that the restored instance would match the properties or the
   origin instance, we will simply check that the user specifies the same
   plan ID.

   Note: Later we might add more complex logic, to allow restore to different
   compatible plans (e.g. from a HA to a non-HA plan).