Implement tests for postgresql_engine.CreateUser()
Implement integration tests to be used with a real Postgresql DB server
for the postgresql_engine implementation.

We add the service postgresql in travis for testing, and all the
boilerplate required. We also document how to run a local docker image.

In order to be able to work with postgres from travis or docker, we must
disable SSL. We add a private attribute to disable requireSSL.
We extended the method SQLEngine.URI() to pass a requireSSL flag.

We also allow override the stateDBName that stores the users to allow
run multiple tests in parallel by generating random databases for the
users state and to access.