Fail restore_snapshot if they provide empty string
Fail if the user passes '-c { "restore_from_latest_snapshot_of": ""}'

To do so we must distinguish between user providing an empty string for
restore_from_latest_snapshot_of or the user not providing the value at
all.

That is achieved by using a pointer to string in the struct that is populated
with the json passed by the user. The pointer will be nil if the user
does not provide that parameter, and empty string if they pass an empty
string.