reassign_owner() trigger ignores superusers
When running the test against a docker or a travis service, we
create the user with a user with superuser permissions. That user
matches the condition of having access to the group role, which makes
the trigger try to change the ownership of the trigger function itself :)

We need to check if the current user is a superuser, and only reassign
the permissions if it is not.