Update tests to support ownership reassign
Update the sqlengine unit tests to work with the implementation
using triggers that change the object ownership.

We create a new superuser that simulates the master user in RDS, to
simulate the behaviour of RDS and also simulate SQL errors by removing
the permissions of that user.

We add the different cases creating users, dropping them, creating
multiple uses for the same database, etc.