User bind/unbind supports legacy bind users
In the previous implementation for postgres we created a unique user
that was returned for every binding, called $dbname+'_owner'.

But in the new implementation we create a user per binding.

In order to support a seamless transition for pre-existing instances
with bindings using the old implementation we add logic that:

 * will not fail on unbinding if the "new style user" does not exist.
   This is because the binding might have been created with the previous
   implementation so the user does not exist.

 * will grant every existing non-superuser user to the new "manager group"
   This group will own all the objects in the new implementation.
   This would allow old bindings to access objects created by new bindings.

 * will reassign all the objects of every existing non-superuser user to
   the new "manager group"
   This would allow the new bindings to access pre-existing objects.

For the transition, we will ask our tenants to rebind the applications.
After a period of grace (e.g. 1 month) we will run a cleanup task to
delete the legacy users and we will revert some of this code.