Use vendoring compatible with Golang 1.8
When switching to the use of Go 1.8 Godep presented the following
warnings when attempting to manage dependencies:

```
godep: WARNING: Godep workspaces (./Godeps/_workspace) are deprecated and support for them will be removed when go1.8 is released.
godep: WARNING: Go version (go1.8) & $GO15VENDOREXPERIMENT= wants to enable the vendor experiment, but disabling because a Godep workspace (Godeps/_workspace) exists
```

The process in the Godep README[1] was followed to move away from using
Godep workspaces.

[1] https://github.com/tools/godep#migrating-to-vendor