Test brokerapi integration for Binding
We want to be sure that the binding REST call to the broker would work
and return the credentials serialised properly.

Specifically we want to test the regresion introduce while upgrading the
brokerapi dependency and described in #20[1]:

> The Bind call now returns a different type for the Credentials hash,
> which no longer serializes to the same JSON document. Specifically, the
> field names are no longer lowercase, which means that several buildpacks
> no longer populate the DATABASE_URL environment variable.

The test will check that the json returned can be deserialised as
expected in a adhoc define type. We add a prefix `The` to all the fields
to be sure that golang json won't deserialise using the field name.

We also check that the json response contain the right strings.

[1] https://github.com/alphagov/paas-rds-broker/pull/20