Remove all the STS api usage
We were using STS to retrieve the userID to be able to generate the ARN
of DBInstances and DBSnapshots. But after upgrading aws-sdk-go we
can retrieve the ARN from the details from the Describe calls, so
we do not need to use STS anymore.

This commit removes all the unnecesary code.