Test that brokerapi returns 410 on lastoperation
We want to be sure that the binding REST call to the broker would work
and return the right response 410 code when calling LastOperation and the
instance does not exist.

Returning 410 Gone, will indicate to the caller that the instance has
been deleted if the last operation was an unbind.

Specifically we want to test the regresion introduced while upgrading the
brokerapi dependency and described in #20[1]:

> Async delete no longer works correctly due to changes in the brokerapi
> library. The frodenas version returns a 410 Gone if the broker
> implementation returns an ErrInstanceDoesNotExist[1], whereas the
> Pivotal-cf version returns a 404 Not Found[2]. The service broker API
> docs specify that a 410 can be returned to indicate to CF that service
> deletion has completed, and this is how this was working previously. A
> 404 is treated as an error, so CF keeps on polling for deletion to
> complete.

The test will check that the status code is 410 when the instance is
missing.

[1] https://github.com/alphagov/paas-rds-broker/pull/20