Remove service.bindable from configuration
The PR from @jmcarp removes the checks for bindable.

It makes sense not to have such logic, as there is no reason for
this broker to return a service that is no bindable. Because that
there is no reason to expose such configuration.

We remove such configuration from the internal broker structures
but we ensure that the value `bindable=true` is returned
by the catalog anyway.