Remove spurios call to PostRestoreTasks in Binding
Binding is calling the method PostRestoreTasks(). It was
added likely during the implementation of that feature accidentally:
the original idea was to implement it in the Binding, but later
was moved to LastOperation.

Calling this function in Binding is a no-op because all the work
is always previously done in LastOperation.

There are also no tests around this call in Binding.