Fix bind and update without custom parameters
A tenant reported that the `cf update-service` command didn't work. I was
able to reproduce it on my development environment:

    ➜  paas-cf git:(master) ✗ cf create-service postgres M-HA-dedicated-9.5 dcarley-test
    …
    ➜  paas-cf git:(master) ✗ cf update-service dcarley-test -p L-HA-dedicated-9.5
    Updating service instance dcarley-test as admin...
    FAILED
    Server error, status code: 502, error code: 10001, message: Service broker error: unexpected end of JSON input

The error occurs when no custom parameters are passed. You can workaround it
by supplying *any* custom parameters e.g. `-c '{"pretty_please":true}'`.
However `-c '{}'` isn't sufficient because the CLI unmarshalls it to an
empty map and then remarshalls the parameters into the request with
`omitempty`.

This worked prior to the dependency updates in 0ac81ce because
`frodenas/brokerapi.UpdateDetails.Parameters` was `map[string]interface{}`
`mapstructure.Decode()` on an empty map was a noop.

The fix is to check whether the parameters are zero-length before attempting
to unmarshall them. We already do this for provision/create-service. It
turned out that bind/bind-service also had the same bug, but we don't deploy
the broker with `rdsbroker.Config.AllowUserBindParameters = true` so the
params are never unmarshalled and we haven't yet been affected.

I've added new unit tests to call out this behaviour for `nil` and `{}`
custom parameters. I'm explicitly setting the "allow" parameter in each
context because I want to ensure that we enter the conditional if we change
the global default for the tests. I considered adding integration tests too,
but I'm not convinced that they would be any clearer or more effective,
since we're interacting with the broker API and quite far removed from what
the CLI or CF may or may not provide us.

It's worth noting that the use of `len(p) > 0` is preferable to `p != nil`
because it works for both an uninitialised field of `json.RawMessage{nil}`
as well as an initialised but empty field of `json.RawMessage{}` or
`json.RawMessage("")`. I don't think they all warrant separate unit tests
though, so I've made one of the second cases the default for all other test
contexts.