Concourse task to run integration tests
Which will be used by alphagov/paas-release-ci. It assumes a single input
directory called "pr" which contains a checkout of a git repo relating to a
pull request.

I've had to do some path hackery in order to run the codebase from GOPATH
otherwise all of the imports will fail. It's not actually possible to move
the input directory, so this method is stolen from Travis CI.

I've used the `governmentpaas/cf-acceptance-tests` container because it has
everything that we need to test this project:

- Go
- Ginkgo
- Godep