Create AWS dependencies within integration tests
The integration tests need an EC2 Security Group and RDS DB Subnet Group in
order to place and connect to RDS instances created by the tests. Previously
these had to be created upfront and their IDs/names passed to the tests as
environment variables.

We want to run these integration tests from alphagov/paas-release-ci but
creating those AWS resources upfront and passing the IDs/names created a
hard dependency between the two repos and makes it difficult to re-use the
pipeline templates for other projects. We did consider using Terraform, but
it required using separate steps in the pipeline with different containers,
which would have again made the interaction between the two repos more
difficult.

Instead we have the test suite create everything that it needs and destroy
them afterwards. As described in the README, this uses the VPC of the
instance that the tests are running from and requires some additional IAM
permissions.

The resources are created with a random UUID suffixed to the name so that
multiple test runs can happen against the same environment. We're not
currently tagging the resources because it requires additional IAM
permissions and API calls, but it should be easy to identify them from their
names.

A separate copy of the config file, with the new resource references, is
written out to a temporary file and used to start the broker.

The `getNetworkMetadata` function does not currently handle cases where an
instance has more than one MAC address, possibly because it's assigned an
Elastic Network Interface. I'm not sure what the output will look because we
don't currently have any of those. I'm pretty confident that the code will
fail in a noisy way though.