Explain how to run MySQL tests locally
When working on the broker being to be able to run the unit tests locally means
that you get much faster feedback on changes, this requires a mysql instance.

We'd prefer to be able to run it in a docker container so we can provide one
set of instructions no matter what OS the workstation is running, but we
haven't been able to make that work, so the instructions are for homebrew.