Add Postgres binding tests
These tests were present in the paas-cf repository[1], but have been
moved here because we want this repository to have tests for integrating
the broker with the AWS API, and we want paas-cf to only test those
integrations which demonstrate the broker works when deployed in our
Cloud Foundry environment. This should produce a faster paas-cf deployment
and enables full integration testing without having to deploy the broker
with Cloud Foundry.

Moving these tests here may mean some commit history is lost, so I shall
re-iterate why they were added initially: we have changed the way Postgres
users are created in the past, and we will be changing it again to make
use of event triggers to transfer the object permissions. Therefore, we
need confidence the user created with a bind has full permissions over the
objecs created by users from previous binds.

[1] https://github.com/alphagov/paas-cf/blob/dee80ec668790d878f995388e71d50d01d13f39f/platform-tests/src/acceptance/rds_broker_test.go#L110-L126