README: Use container for local MySQL tests
I think this is preferable to using Homebrew because it:

- doesn't pollute my main machine with necessary dependencies
- is consistent with our documentation for Postgres

Unfortunately we have to wait for MySQL to have finished starting before we
can use it.

I've adjusted the instructions to suit Docker for Mac and Docker for Linux,
which don't need you to find the IP of the Docker VM or manually configure
port forwarding. We don't need to specify the hostname environment variables
because the default of `localhost` should work for everyone.

Finally, the containers can be stopped and deleted in one command.