Test connection with new user
At the moment, the user and password created, are not tested when
establishing the connection. To prevent issues down the line, we'd like
to cover that case with some tests.

It has pointed out the issues in our code, where we didn't use back
ticks to identify the database in a query related to GRANT.

It is caused, by the fact that not all characters we are generating,
particularly the hyphen, are not acceptable in an unquoted identirfier.

Ref: https://dev.mysql.com/doc/refman/5.5/en/identifiers.html