Refactor MigrateLegacyUser to support RDS Postgres
This commit refactors the MigrateLegacyUser fuction after an update from AWS via a support ticket.

As the user we have is not a true Superuser it is essential to add this
user to the new role before performing the reassign owner statement. At
the end of this function we revoke this role from our Superuser.