Use parallelism for integration tests
Since we added the MySQL integration tests it now takes 1h30m for them to
run in Concourse. Attempt to reduce the time we have to wait by running 8
nodes at once.

This shouldn't cause any more RDS instances to be created than previous
because of the way that we have structured our tests to not have multiple
`It` statements (which could be run on different nodes) that depend on a
service instance being provisioned in `BeforeEach`.