MySQL integration test connect + rebind perms
To match what we do for Postgres.

Unfortunately the Go driver for MySQL requires a completely different
connection string to the URL that the broker returns. So we have to
reconstruct it from the parsed URL. We also ignore the query params because
it doesn't support the `reconnect` param which we return by default.