Refactor integration tests to use shared examples
We now have the same integration tests for Postgres and MySQL. So we can use
shared examples to de-duplicate the test code, which makes it easier to read
and maintain:

- https://onsi.github.io/ginkgo/#shared-example-patterns

I double checked that the tests were the same by extracting them to separate
files and diffing them. The only differences was the service name in the
`It` description, which I've omitted from the shared examples.

The diff is easier to read with `git show -b` to ignore whitespace.