Simplify catalog for integration tests
The generic service and plan IDs (e.g. "Service-1" and "Plan-1") made it
hard for me to understand whether a particular test was using Postgres or
MySQL without cross-referencing the catalog. I also overlooked the subtlety
that the standard Provision/Bind/Deprovision tests for Postgres were using
`skip_final_snapshot`, presumably to save time and cost, but the MySQL tests
weren't.

So I've restructured them to:

- only have two services, Postgres and MySQL, each with more than one plan,
  like we do in real environments
- have names that reflect what they are
- disable final snapshots for the standard MySQL tests
- use `t2.micro` throughout because it's cheaper and we're not doing
  anything very intensive

I was struggling to see the value of testing every attribute in the catalog,
so I've removed the assertions for the individual plans, as well as some
duplicate assertions on the service ID.