Force TLS for MySQL service instances
We want tenants to always use TLS when connecting to MySQL service instances
because it protects data in-transit over the network and is good practice.

MySQL enforces this on a per-user basis, unlike Postgres which enforces it
on a per-server basis, so this has to be done as part of the user creation
during bind.

We will add an acceptance test to check whether TLS is possible and non-TLS
is rejected to alphagov/paas-cf, rather than this repo, to match the same
test that we have for Postgres.

This used to be a `GRANT` option but it has been deprecated in MySQL 5.7
(which we specify in the catalog) for a `CREATE USER` option:

- https://dev.mysql.com/doc/refman/5.7/en/create-user.html#create-user-tls
- https://dev.mysql.com/doc/refman/5.7/en/grant.html#grant-other-characteristics

There appears to be no standardisation between libraries of various
languages of the query params in the URI returned in `credentials.uri`. So
we've decided not to modify it for the moment. We might need to revisit it
in the future based on feedback from users.

The JDBC URI returned in `credentials.jdbcuri` does appear to have better
standardisation, at least according to this documentation:

- https://dev.mysql.com/doc/connector-j/5.1/en/connector-j-reference-configuration-properties.html

However I wasn't able to test the JDBC URI because all of the Java applications I
tested appeared to use `DATABASE_URL` based on `credentials.uri`.
Nonetheless, I don't think we should modify it because:

- none of the Java apps I tested needed any further configuration to use TLS
  by default
- the documentation says that `useSSL` and `verifyServerCertificate` will
  default to `true` when connecting to MySQL 5.7.6+