Utility to get master password for an instance
This can be useful when debugging a service instance.

It can be used with the username from:

    aws rds describe-db-instances \
      --db-instance-identifier <DB_PREFIX>-<SERVICE_INSTANCE_ID> \
      --query 'DBInstances[].MasterUsername'

I've made the username and password length constants public so that they can
be re-used by this separate package and not get out-of-sync.

I've made it interactive, rather than taking CLI arguments, to prevent
someone from storing the master seed password in their shell history.