Fix snapshot restore with legacy users.
As part of the post-restore process for a database instance created from
a snapshot, the broker drops all existing users (not groups) that exist.
This is because their corresponding bindings won't exist in CF.

Postgres doesn't allow dropping a user that owns any objects, or has any
permissions assigned (other than membership of other roles). The legacy
user had been granted permissions on the database, which resulted in the
post-migration tasks failing with the error:

ERROR:  role "rdsbroker_<UUID>_owner" cannot be dropped because some objects depend on it
DETAIL:  privileges for database rdsbroker_<UUID>

Updating MigrateLegacyAdminUsers to revoke these permissions from the
legacy user allows the user to be dropped. The legacy user no longer
needs these permissions because it's now been made a member of the
_manager group, and gets its permissions via that.