Revert ownership of public schema
We currently have functions and triggers (reassigned_owned) installed
into the "public" schema.

Making the user the owner of the public schema enables them to
drop and cascadingly delete everything within the schema including any
funcs/triggers installed by the broker to manage object ownership.

We decided that it is not wise at this time to give full priviledges to
the public schema.