Create Postgres extensions
We want users of Postgres instances to be able to use popular extensions[1].

Given there is little or no risk in enabling unneeded extensions we can
simplify the experience for users by doing it for them.

The options we explored:

* enabling them at the AWS level. This is not possible, as database parameter
groups do not allow you to configure extensions in any way.
* enabling them on `bind` operations. This ensures the instance is ready to
have these SQL statements run, but conceptually does not fit, as it is nothing
to do with the binding process.
* using the last operation polling: by waiting for the instance to be ready
and for all post-restore tasks (from snapshots), to be completed we ensure the
instance is in a state to have the necessary SQL statements run. It also ensures
the instance isn't used by users until the extensions are created.

We went with the latter option.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/148570883