Resolve race condition in RDS broker startup
Prior to this commit the broker prints that it is ready to
accept connections before it has even bound to the TCP port.
This means that something trying to connect to its HTTP server
as soon as the message is printed would be in a race condition
with the port binding. In addition if that binding fails the
'ready' message would still have been printed.

@henrytk and I suspect this is responsible for a failure in the
`rds-broker` pipeline [1]. The integrations tests start running
as soon as the broker prints that it is ready, and they failed
saying `connection refused` on the port the broker listens to.
Given the transitory nature of this failure, this race condition
seems a likely cause. Even if not there is the potential for
failure here.

From this commit we now listen to the port before printing the
'ready' message. This should resolve the problem.

[1] https://concourse.build.ci.cloudpipeline.digital/teams/main/pipelines/rds-broker/jobs/tag-releases/builds/4