Invert the use of a parameter
We'd like to make it less confusing to our tenants, to run the update on
their RDS instances. Right now, the default behaviour is to wait for a
maintenance window in order to apply any changes. With this change, the
default will be to apply the changes immediately but give the users an
option to wait for the maintenance window if they wish to do so.